rename genericRoller folder to roller
