+++ src/main/java/frc/robot/subsystems/roller/RollerIOReal.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems.genericRoller;
+package frc.robot.subsystems.roller;
 
 import static edu.wpi.first.units.Units.Amps;
 import static edu.wpi.first.units.Units.Celsius;
@@ -23,6 +23,8 @@ import edu.wpi.first.units.measure.AngularVelocity;
 import edu.wpi.first.units.measure.Current;
 import edu.wpi.first.units.measure.Temperature;
 import edu.wpi.first.units.measure.Voltage;
+import frc.robot.subsystems.genericRoller.RollerIOInputsAutoLogged;
+
 import java.util.Optional;
 import java.util.function.Consumer;
